rest_api/requirements.txt
asgiref==3.4.1
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
cloud-init==22.3.4
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
distlib==0.3.6
distro-info===0.18ubuntu0.18.04.1
Django==3.2.16
filelock==3.4.1
gunicorn==20.1.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
importlib-metadata==4.8.3
importlib-resources==5.4.0
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
language-selector==0.1
MarkupSafe==1.0
netifaces==0.10.4
oauthlib==2.0.6
PAM==0.4.2
pexpect==4.2.1
platformdirs==2.4.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
PyGObject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.7
python-debian==0.1.32
pytz==2022.6
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
service-identity==16.0.0
six==1.11.0
sos==4.4
sqlparse==0.4.3
ssh-import-id==5.7
systemd-python==234
Twisted==17.9.0
typing_extensions==4.1.1
ubuntu-advantage-tools==27.11.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
virtualenv==20.16.6
zipp==3.6.0
zope.interface==4.3.2
